security: resolve code-scanning alerts (CodeQL C/C++ + OSSF Scorecard hardening)
CodeQL C/C++:
- src/fs/fsck/fsck.c: promote imap/bmap bitmap size math to size_t before the
  multiply (cpp/integer-multiplication-cast-to-long x2) -- imap_blocks/bmap_blocks
  and dev->block_size are uint32_t, so the product could overflow 32-bit before
  the size_t conversion passed to kmalloc.
- .github/codeql/codeql-config.yml (new) + codeql.yml config-file: exclude
  third_party/ from analysis. The 10 cpp/weak-cryptographic-algorithm alerts are
  all vendored BearSSL DES/3DES-CBC files (protocol completeness); CapyOS never
  selects DES cipher suites and vendored code is not fixed in-tree.

OSSF Scorecard hardening:
- Pin all 25 GitHub Actions `uses:` to full commit SHAs (with version comments)
  across the 5 workflows (PinnedDependenciesID).
- Add top-level least-privilege `permissions: contents: read` to codeql.yml
  (TokenPermissionsID); release-artifacts.yml already scopes contents:write to
  the job with a top-level read default.
- Add SECURITY.md (SecurityPolicyID) with private advisory reporting.

Docs: compatibility-matrix pins CapyUI 2.22.7 (input bridge Slice 7.5; the
gfx-owned redirect test suite requires >=2.22.7).

Not addressed (repo-governance, owner decision; branch protection would block the
current direct-push release flow): BranchProtectionID, CodeReviewID,
CIIBestPracticesID.

Validated: workflow YAML parses; make all64 PROFILE=full
CAPYOS_DESKTOP_GRAPHICAL_BROWSER=1 TOOLCHAIN64=host clean (fsck.o rebuilt).

+++ src/fs/fsck/fsck.c
@@ -92,7 +92,7 @@ int fsck_check(struct block_device *dev, struct fsck_result *result) {
 
   /* Read inode bitmap from disk */
   uint32_t imap_blocks = (inode_bytes + dev->block_size - 1) / dev->block_size;
-  uint8_t *disk_imap = (uint8_t *)kmalloc(imap_blocks * dev->block_size);
+  uint8_t *disk_imap = (uint8_t *)kmalloc((size_t)imap_blocks * dev->block_size);
   if (disk_imap) {
     for (uint32_t i = 0; i < imap_blocks; i++)
       fsck_read_block(dev, imap_start + i, disk_imap + i * dev->block_size);
@@ -100,7 +100,7 @@ int fsck_check(struct block_device *dev, struct fsck_result *result) {
 
   /* Read block bitmap from disk */
   uint32_t bmap_blocks = (block_bytes + dev->block_size - 1) / dev->block_size;
-  uint8_t *disk_bmap = (uint8_t *)kmalloc(bmap_blocks * dev->block_size);
+  uint8_t *disk_bmap = (uint8_t *)kmalloc((size_t)bmap_blocks * dev->block_size);
   if (disk_bmap) {
     for (uint32_t i = 0; i < bmap_blocks; i++)
       fsck_read_block(dev, bmap_start + i, disk_bmap + i * dev->block_size);
